// src/app/models/academics/student-details-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface StudentDetailsView {
    schoolBoardId?: number;
    classId?: number;
    sectionId?: number;
    rollNumberSortBy?: Array<string>;
    sortOrder?: number;
    pageNumber?: number;
    pageSize?: number;

}