// src/app/models/academics/roll-number-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * RollNumberView
 */
export interface RollNumberView { 

    /**
     * Gets or sets  ClassId
     */
    classId?: number;
    /**
     * Gets or sets  SectionId
     */
    sectionId?: number;
    /**
     * Gets or sets  IsRearrange
     */
    isRearrange?: number;
    /**
     * Gets or sets  StartingNumber
     */
    startingNumber?: number;
     /**
     * Gets or sets  rollNumberSortBy
     */

    rollNumberSortBy?: Array<string>;
}